+++ packages/ossa-demo/src/components/picker/demo/index.tsx
@@ -22,10 +22,19 @@ const initialListApi = {
   head: ["参数", "说明", "类型", "默认值"],
   data: [
     {
-      list: ["range", "取值范围，必选", "Array<number|string>", "-"],
+      list: ["title", "标题，可选", "string", "-"],
     },
     {
-      list: ["value", "取值索引，必选", "number", "0"],
+      list: ["isShow", "是否显示，传入布尔值可手动控制，可选", "boolean", "-"],
+    },
+    {
+      list: ["mode", "类型，mode 为 selector 时为一维数组，mode 为 multiSelector 时为二维数组，可选", "'selector'|'multiSelector'", "'selector'"],
+    },
+    {
+      list: ["range", "取值范围，必选", "Array<number|string> | Array<Array<number|string>", "-"],
+    },
+    {
+      list: ["value", "取值索引，必选", "number|number[]", "0"],
     },
   ],
 };
@@ -34,22 +43,31 @@ const initialListEvent = {
   head: ["函数名", "说明", "参数"],
   data: [
     {
-      list: ["onConfirm", "点击确认按钮时触发，可选", "-"],
+      list: ["onConfirm", "点击确认按钮时触发，可选", "(value) => void"],
     },
     {
-      list: ["onCancel", "点击取消按钮时触发，可选", "-"],
+      list: ["onCancel", "点击取消按钮时触发，可选", "() => void"],
+    },
+    {
+      list: ["onChange", "值改变时触发，可选", "(value) => void"],
     },
   ],
 };
 
 const range = ["中国", "美国", "日本", "泰国", "印度"];
 
+const multiRange = [
+  ["星期一", "星期二", "星期三", "星期四", "星期五", "星期六", "星期日"],
+  ["上午", "下午", "晚上"],
+]
+
 export default function Index(props: Props) {
   const [listApi] = useState(initialListApi);
 
   const [listEvent] = useState(initialListEvent);
   const classObject = getClassObject(); //组件修饰
-  const [vNormal, setVNormal] = useState(2);
+  const [vNormal, setVNormal] = useState(0);
+  const [vMulti, setVMulti] = useState([0, 0]);
   const onConfirmNormal = (value) => {
     setVNormal(Number(value));
   };
@@ -72,8 +90,37 @@ export default function Index(props: Props) {
               console.log("cancel");
             }}
             onConfirm={onConfirmNormal}
+            onChange={(value) => {
+              console.log(value);
+            }}
           >
             <OsList title='国籍' desc={range[vNormal]}></OsList>
+          </OsPicker>
+        </View>
+      </DemoBlock>
+      <DemoBlock title='多列'>
+        <View className='block-section'>
+          <OsPicker
+            title='时间'
+            mode='multiSelector'
+            range={multiRange}
+            value={vMulti}
+            onCancel={() => {
+              console.log("cancel");
+            }}
+            onConfirm={(value) => {
+              console.log(value);
+              setVMulti(value as number[])
+            }}
+            onChange={(value) => {
+              console.log(value);
+              if (value[0] !== vMulti[0]) {
+                value[1] = 0;
+              }
+              setVMulti(value as number[]);
+            }}
+          >
+            <OsList title='时间' desc={multiRange.map((rang, index) => rang[vMulti[index]]).join(',')}></OsList>
           </OsPicker>
         </View>
       </DemoBlock>