+++ packages/ossa/types/picker.d.ts
@@ -6,18 +6,34 @@ import { OsCommonValueChangeCallback, OsCommonFunction } from "./common";
  */
 interface PickerProps extends OsComponent {
   /**
-   * 取值范围
+   * 选择器类型，默认为 selector
    */
-  range: Array<string>;
+  mode?: "selector" | "multiSelector";
   /**
-   * 取值索引
+   * 是否显示，传入布尔值可手动控制，可选
    */
-  value: number;
+  isShow?: boolean;
+  /**
+   * 标题
+   */
+  title?: string;
+  /**
+   * 类型，mode 为 selector 时为一维数组，mode 为 multiSelector 时为二维数组，默认为 selector
+   */
+  range: Array<number|string> | Array<Array<number|string>>;
+  /**
+   * 取值索引，mode 为 selector 时为 number，mode 为 multiSelector 时为数组
+   */
+  value: number | number[];
+  /**
+   * 所选项变化时触发
+   */
+  onChange?: OsCommonValueChangeCallback<number | string | Array<number|string>>;
   /**
    * 确认选择
    * @param {number | string} v 当前选中的值
    */
-  onConfirm?: OsCommonValueChangeCallback<number | string>;
+  onConfirm?: OsCommonValueChangeCallback<number | string | Array<number|string>>;
   /**
    * 取消选择
    */