+++ packages/ossa-demo/CHANGELOG.md
@@ -1,5 +1,15 @@
 # ossa-demo
 
+## 1.9.0-beta.0
+
+### Minor Changes
+
+- [`9712e8e`](https://github.com/NeteaseYanxuan/OSSA/commit/9712e8e1de8c62e165d96b8d7e3e1ab42bcd6311) Thanks [@CaedmonW](https://github.com/CaedmonW)! - Picker support multi
+
+### Patch Changes
+
+- Updated dependencies [[`9712e8e`](https://github.com/NeteaseYanxuan/OSSA/commit/9712e8e1de8c62e165d96b8d7e3e1ab42bcd6311)]:
+  - ossaui@0.7.0-beta.0
 ## 1.8.3
 
 ### Patch Changes